+++ modules/AdminToolbox.Office365/ChangeLog.md
@@ -31,6 +31,7 @@
 * **2.3.0.1** Fixed for invalid comment character in function Get-TenantDiagrams
 * **2.4.0.0** Added Markdown Documentation and link help
 * **2.5.0** Update versioning standard and added LicenseURI.
+* **2.6.0** Added Clear-TeamsCache
 
 #The line below this one is used strictly by a github action. That action is responsible for publishing new versions of this module to Powershell Gallery, upon pushing an iterative version to Github.
-2.5.0
+2.6.0