+++ modules/AdminToolbox/ChangeLog.md
@@ -154,6 +154,7 @@
 * **9.4.0** Update required modules
 * **9.5.0** Update required modules
 * **9.6.0** Update required modules
+* **9.7.0** Update required modules
 
 #The line below this one is used strictly by a github action. That action is responsible for publishing new versions of this module to Powershell Gallery, upon pushing an iterative version to Github.
-9.6.0
+9.7.0